# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "GH Trust MFB API",
  "description": "GH Trust International Ltd \u2014 Microfinance Banking API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/auth/register/bvn": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Register with BVN",
    "description": "Submit BVN only. We verify via Dojah BVN Advanced, create your profile from registry data, and send OTP to the phone linked to your BVN.",
    "operationId": "register_with_bvn_api_v1_auth_register_bvn_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BvnRegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__modules__auth__schemas__OtpSentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/register/verify-otp": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Verify registration OTP",
    "description": "With selfie checks on (the default) returns `status: selfie_required` and a `registration_token` for `/auth/register/selfie`; otherwise opens the account and signs the device in. Send `device` from mobile clients.",
    "operationId": "verify_registration_otp_api_v1_auth_register_verify_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VerifyRegistrationOtpRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/AuthTokenResponse"
          },
          {
           "$ref": "#/components/schemas/SelfieRequiredResponse"
          }
         ],
         "title": "Response Verify Registration Otp Api V1 Auth Register Verify Otp Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/register/selfie": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Match a selfie to the BVN photo and open the account",
    "description": "Dojah compares the selfie with the BVN photo. A match opens the account and signs the device in; `SELFIE_NO_MATCH` includes `attempts_left`; after the last attempt (`SELFIE_ATTEMPTS_EXCEEDED`) the customer must visit a branch.",
    "operationId": "verify_registration_selfie_api_v1_auth_register_selfie_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegistrationSelfieRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthTokenResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/register/resend-otp": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Resend registration OTP",
    "operationId": "resend_registration_otp_api_v1_auth_register_resend_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResendRegistrationOtpRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__modules__auth__schemas__OtpSentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/login/request-otp": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Request login OTP",
    "operationId": "request_login_otp_api_v1_auth_login_request_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneLoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__modules__auth__schemas__OtpSentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/login/verify-otp": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Verify login OTP",
    "description": "Signs the phone in (`status: signed_in`), unless it's a new phone and another one is signed in: then `status: approval_required` and the sign-in waits for the customer to approve it on the other phone (see `/auth/device-approvals`).",
    "operationId": "verify_login_otp_api_v1_auth_login_verify_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VerifyLoginOtpRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/AuthTokenResponse"
          },
          {
           "$ref": "#/components/schemas/DeviceApprovalRequiredResponse"
          }
         ],
         "title": "Response Verify Login Otp Api V1 Auth Login Verify Otp Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/login/resend-otp": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Resend login OTP",
    "operationId": "resend_login_otp_api_v1_auth_login_resend_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneLoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__modules__auth__schemas__OtpSentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/token/refresh": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Refresh access token",
    "description": "Exchange a refresh token for a new access + refresh pair. The refresh token rotates: store the new one and discard the old. Presenting an old refresh token again signs the session out (`REFRESH_TOKEN_REUSED`). Clients should run at most one refresh at a time.",
    "operationId": "refresh_token_api_v1_auth_token_refresh_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RefreshTokenRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenPair"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Sign out this device",
    "operationId": "logout_api_v1_auth_logout_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "forget_device",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Also stop trusting this phone (\"Not you?\"): next time needs an SMS code.",
       "default": false,
       "title": "Forget Device"
      },
      "description": "Also stop trusting this phone (\"Not you?\"): next time needs an SMS code."
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/logout-all": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Sign out every device",
    "operationId": "logout_all_api_v1_auth_logout_all_post",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/auth/sessions": {
   "get": {
    "tags": [
     "Authentication"
    ],
    "summary": "Signed-in devices",
    "operationId": "list_sessions_api_v1_auth_sessions_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SessionResponse"
         },
         "type": "array",
         "title": "Response List Sessions Api V1 Auth Sessions Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/auth/sessions/{session_id}": {
   "delete": {
    "tags": [
     "Authentication"
    ],
    "summary": "Sign out a specific device",
    "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "session_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Session Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/me": {
   "get": {
    "tags": [
     "Authentication"
    ],
    "summary": "Current customer profile",
    "operationId": "get_current_profile_api_v1_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CustomerProfileResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
  "/api/v1/admin/auth/login/request-otp": {
   "post": {
    "tags": [
     "Admin"
    ],
    "summary": "Staff login \u2014 request OTP",
    "operationId": "staff_request_login_otp_api_v1_admin_auth_login_request_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StaffLoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__modules__admin__schemas__OtpSentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/login/verify-otp": {
   "post": {
    "tags": [
     "Admin"
    ],
    "summary": "Staff login \u2014 verify OTP",
    "operationId": "staff_verify_login_otp_api_v1_admin_auth_login_verify_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VerifyStaffOtpRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StaffAuthTokenResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/login/resend-otp": {
   "post": {
    "tags": [
     "Admin"
    ],
    "summary": "Staff login \u2014 resend OTP",
    "operationId": "staff_resend_login_otp_api_v1_admin_auth_login_resend_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StaffLoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__modules__admin__schemas__OtpSentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/token/refresh": {
   "post": {
    "tags": [
     "Admin"
    ],
    "summary": "Staff \u2014 refresh access token",
    "description": "Send the refresh token in the body, or (web portal) rely on the httpOnly cookie.",
    "operationId": "staff_refresh_token_api_v1_admin_auth_token_refresh_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RefreshTokenRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Payload"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StaffTokenPair"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/logout": {
   "post": {
    "tags": [
     "Admin"
    ],
    "summary": "Staff \u2014 sign out",
    "description": "Ends the session identified by the bearer token and/or the refresh cookie. Works even when the access token has already expired.",
    "operationId": "staff_logout_api_v1_admin_auth_logout_post",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/admin/auth/me": {
   "get": {
    "tags": [
     "Admin"
    ],
    "summary": "Current staff profile",
    "operationId": "staff_me_api_v1_admin_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StaffResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "patch": {
    "tags": [
     "Admin"
    ],
    "summary": "Update own profile",
    "description": "Name, job title and avatar colour. Send an empty job title to clear it.",
    "operationId": "update_staff_me_api_v1_admin_auth_me_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StaffSelfUpdateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StaffResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AuthTokenResponse": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In",
      "description": "Access token lifetime in seconds"
     },
     "refresh_token": {
      "type": "string",
      "title": "Refresh Token",
      "description": "Opaque; rotates on every refresh. Store securely."
     },
     "refresh_expires_in": {
      "type": "integer",
      "title": "Refresh Expires In",
      "description": "Refresh token lifetime in seconds"
     },
     "session_id": {
      "type": "string",
      "title": "Session Id"
     },
     "status": {
      "type": "string",
      "const": "signed_in",
      "title": "Status",
      "default": "signed_in"
     },
     "customer": {
      "$ref": "#/components/schemas/CustomerProfileResponse"
     },
     "device_token": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Token",
      "description": "Store securely on this phone and send it in `device` next time: it lets the customer sign back in here with their PIN, and skips new-device approval."
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "refresh_token",
     "refresh_expires_in",
     "session_id",
     "customer"
    ],
    "title": "AuthTokenResponse"
   },
...
   "BvnRegisterRequest": {
    "properties": {
     "bvn": {
      "type": "string",
      "maxLength": 11,
      "minLength": 11,
      "title": "Bvn",
      "description": "11-digit Bank Verification Number"
     }
    },
    "type": "object",
    "required": [
     "bvn"
    ],
    "title": "BvnRegisterRequest"
   },
...
   "DeviceApprovalRequiredResponse": {
    "properties": {
     "status": {
      "type": "string",
      "const": "approval_required",
      "title": "Status",
      "default": "approval_required"
     },
     "approval_id": {
      "type": "string",
      "title": "Approval Id"
     },
     "approval_secret": {
      "type": "string",
      "title": "Approval Secret",
      "description": "Keep in memory; proves this phone started the sign-in."
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "approver_devices": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Approver Devices",
      "description": "Names of the signed-in phones that can approve"
     },
     "fallback_needs_pin": {
      "type": "boolean",
      "title": "Fallback Needs Pin",
      "description": "The lost-phone route asks for the sign-in PIN as well as the BVN."
     }
    },
    "type": "object",
    "required": [
     "approval_id",
     "approval_secret",
     "expires_in",
     "approver_devices",
     "fallback_needs_pin"
    ],
    "title": "DeviceApprovalRequiredResponse",
    "description": "Login OTP was right, but this is a new phone: confirm it on a signed-in one first."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PhoneLoginRequest": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 15,
      "minLength": 10,
      "title": "Phone",
      "description": "Phone number linked to BVN account"
     }
    },
    "type": "object",
    "required": [
     "phone"
    ],
    "title": "PhoneLoginRequest"
   },
...
   "RefreshTokenRequest": {
    "properties": {
     "refresh_token": {
      "type": "string",
      "maxLength": 256,
      "minLength": 20,
      "title": "Refresh Token"
     }
    },
    "type": "object",
    "required": [
     "refresh_token"
    ],
    "title": "RefreshTokenRequest"
   },
   "RegistrationSelfieRequest": {
    "properties": {
     "registration_token": {
      "type": "string",
      "maxLength": 128,
      "minLength": 20,
      "title": "Registration Token"
     },
     "selfie_image": {
      "type": "string",
      "maxLength": 4000000,
      "minLength": 1000,
      "title": "Selfie Image",
      "description": "JPEG/PNG, base64 (a data: URL prefix is accepted and removed). Face the camera in good light."
     },
     "device": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DeviceInfo"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "registration_token",
     "selfie_image"
    ],
    "title": "RegistrationSelfieRequest"
   },
...
   "ResendRegistrationOtpRequest": {
    "properties": {
     "bvn": {
      "type": "string",
      "maxLength": 11,
      "minLength": 11,
      "title": "Bvn"
     }
    },
    "type": "object",
    "required": [
     "bvn"
    ],
    "title": "ResendRegistrationOtpRequest"
   },
...
   "SelfieRequiredResponse": {
    "properties": {
     "status": {
      "type": "string",
      "const": "selfie_required",
      "title": "Status",
      "default": "selfie_required"
     },
     "registration_token": {
      "type": "string",
      "title": "Registration Token",
      "description": "Send with the selfie. Single use; keep in memory only."
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "attempts_left": {
      "type": "integer",
      "title": "Attempts Left"
     },
     "first_name": {
      "type": "string",
      "title": "First Name"
     }
    },
    "type": "object",
    "required": [
     "registration_token",
     "expires_in",
     "attempts_left",
     "first_name"
    ],
    "title": "SelfieRequiredResponse",
    "description": "SMS code accepted; the account opens once a selfie matches the BVN photo."
   },
   "SessionResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "device_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Name"
     },
     "platform": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Platform"
     },
     "app_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "App Version"
     },
     "ip_address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ip Address"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_used_at": {
      "type": "string",
      "format": "date-time",
      "title": "Last Used At"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     },
     "current": {
      "type": "boolean",
      "title": "Current",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "last_used_at",
     "expires_at"
    ],
    "title": "SessionResponse"
   },
...
   "TokenPair": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In",
      "description": "Access token lifetime in seconds"
     },
     "refresh_token": {
      "type": "string",
      "title": "Refresh Token",
      "description": "Opaque; rotates on every refresh. Store securely."
     },
     "refresh_expires_in": {
      "type": "integer",
      "title": "Refresh Expires In",
      "description": "Refresh token lifetime in seconds"
     },
     "session_id": {
      "type": "string",
      "title": "Session Id"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "refresh_token",
     "refresh_expires_in",
     "session_id"
    ],
    "title": "TokenPair"
   },
...
   "VerifyLoginOtpRequest": {
    "properties": {
     "otp": {
      "type": "string",
      "maxLength": 8,
      "minLength": 4,
      "title": "Otp"
     },
     "device": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DeviceInfo"
       },
       {
        "type": "null"
       }
      ]
     },
     "phone": {
      "type": "string",
      "maxLength": 15,
      "minLength": 10,
      "title": "Phone"
     }
    },
    "type": "object",
    "required": [
     "otp",
     "phone"
    ],
    "title": "VerifyLoginOtpRequest"
   },
...
   "VerifyRegistrationOtpRequest": {
    "properties": {
     "otp": {
      "type": "string",
      "maxLength": 8,
      "minLength": 4,
      "title": "Otp"
     },
     "device": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DeviceInfo"
       },
       {
        "type": "null"
       }
      ]
     },
     "bvn": {
      "type": "string",
      "maxLength": 11,
      "minLength": 11,
      "title": "Bvn"
     }
    },
    "type": "object",
    "required": [
     "otp",
     "bvn"
    ],
    "title": "VerifyRegistrationOtpRequest"
   },
...
   "app__modules__auth__schemas__OtpSentResponse": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     },
     "phone_masked": {
      "type": "string",
      "title": "Phone Masked"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "purpose": {
      "type": "string",
      "title": "Purpose"
     },
     "dev_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dev Code",
      "description": "Local development with mocked SMS only: the code, so a dev build can fill it in. Never present when SMS is real or outside APP_ENV=development."
     }
    },
    "type": "object",
    "required": [
     "message",
     "phone_masked",
     "expires_in",
     "purpose"
    ],
    "title": "OtpSentResponse"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}